lanms create cmake library
create lanms.cpp to seperate source from header
test code main.cpp
eval.py fix typo
