+++ eval.py
@@ -68,15 +68,15 @@ def resize_image(im, max_side_len=2400):
     return im, (ratio_h, ratio_w)
 
 
-def detect(score_map, geo_map, timer, score_map_thresh=0.8, box_thresh=0.1, nms_thres=0.2):
+def detect(score_map, geo_map, timer, score_map_thresh=0.8, box_thresh=0.1, nms_thresh=0.2):
     '''
     restore text boxes from score map and geo map
     :param score_map:
     :param geo_map:
     :param timer:
-    :param score_map_thresh: threshhold for score map
-    :param box_thresh: threshhold for boxes
-    :param nms_thres: threshold for nms
+    :param score_map_thresh: threshold for score map
+    :param box_thresh: threshold for boxes
+    :param nms_thresh: threshold for nms
     :return:
     '''
     if len(score_map.shape) == 4:
@@ -96,8 +96,8 @@ def detect(score_map, geo_map, timer, score_map_thresh=0.8, box_thresh=0.1, nms_
     timer['restore'] = time.time() - start
     # nms part
     start = time.time()
-    # boxes = nms_locality.nms_locality(boxes.astype(np.float64), nms_thres)
-    boxes = lanms.merge_quadrangle_n9(boxes.astype('float32'), nms_thres)
+    # boxes = nms_locality.nms_locality(boxes.astype(np.float64), nms_thresh)
+    boxes = lanms.merge_quadrangle_n9(boxes.astype('float32'), nms_thresh)
     timer['nms'] = time.time() - start
 
     if boxes.shape[0] == 0:
