+++ lanms/.gitignore
@@ -1,1 +1,2 @@
 adaptor.so
+build
